+++ CHANGES.md
@@ -35,8 +35,8 @@ Beta Releases
 * Added `BoundingSphere.projectTo2D`.
 * Added `Extent.fromDegrees`.
 * Added `czm_tangentToEyeSpaceMatrix` built-in GLSL function.
-* Added `extrudedOptions` parameters to `ExtentGeometry`, in which height, closeTop and closeBottom can be specified
 * Improved the performance of drawing polygons created with `configureFromPolygonHierarchy`.
+* Added extrusion to `ExtentGeometry`
 
 ### b18 - 2013-07-01